feat(shared-charts): abstract shared rollup config for charts
