+++ packages/shared/charts/rollup/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "../dist/rollup/index.js"
+}
